move TestLoaderFiltered to test.framework.utilities
