+++ test/framework/parallelbuild.py
@@ -31,7 +31,7 @@ import os
 import re
 import stat
 import sys
-from test.framework.utilities import EnhancedTestCase, init_config
+from test.framework.utilities import EnhancedTestCase, TestLoaderFiltered, init_config
 from unittest import TextTestRunner
 from vsc.utils.fancylogger import setLogLevelDebug, logToScreen
 
@@ -42,7 +42,6 @@ from easybuild.tools.job import pbs_python
 from easybuild.tools.job.pbs_python import PbsPython
 from easybuild.tools.parallelbuild import build_easyconfigs_in_parallel, submit_jobs
 from easybuild.tools.robot import resolve_dependencies
-from easybuild.tools.testfilter import TestLoaderFiltered
 
 
 # test GC3Pie configuration with large resource specs
