+++ test/framework/run.py
@@ -33,7 +33,7 @@ import os
 import re
 import signal
 import sys
-from test.framework.utilities import EnhancedTestCase, init_config
+from test.framework.utilities import EnhancedTestCase, TestLoaderFiltered, init_config
 from unittest import TextTestRunner
 from vsc.utils.fancylogger import setLogLevelDebug, logToScreen
 
@@ -41,7 +41,6 @@ from easybuild.tools.build_log import EasyBuildError
 from easybuild.tools.filetools import read_file
 from easybuild.tools.run import run_cmd, run_cmd_qa, parse_log_for_error
 from easybuild.tools.run import _log as run_log
-from easybuild.tools.testfilter import TestLoaderFiltered
 
 
 class RunTest(EnhancedTestCase):
