+++ test/framework/toolchain.py
@@ -35,7 +35,7 @@ import sys
 import tempfile
 from distutils.version import LooseVersion
 from unittest import TextTestRunner
-from test.framework.utilities import EnhancedTestCase, find_full_path, init_config
+from test.framework.utilities import EnhancedTestCase, TestLoaderFiltered, find_full_path, init_config
 
 import easybuild.tools.build_log
 import easybuild.tools.modules as modules
@@ -44,7 +44,6 @@ from easybuild.framework.easyconfig.easyconfig import EasyConfig, ActiveMNS
 from easybuild.tools import systemtools as st
 from easybuild.tools.build_log import EasyBuildError
 from easybuild.tools.filetools import write_file
-from easybuild.tools.testfilter import TestLoaderFiltered
 from easybuild.tools.toolchain.utilities import get_toolchain, search_toolchain
 
 easybuild.tools.toolchain.compiler.systemtools.get_compiler_family = lambda: st.POWER
