+++ test/framework/toolchainvariables.py
@@ -30,10 +30,9 @@ Unit tests for tools/toolchain/variables.py.
 """
 import sys
 
-from test.framework.utilities import EnhancedTestCase
+from test.framework.utilities import EnhancedTestCase, TestLoaderFiltered
 from unittest import TextTestRunner
 
-from easybuild.tools.testfilter import TestLoaderFiltered
 from easybuild.tools.toolchain.toolchainvariables import ToolchainVariables
 from easybuild.tools.toolchain.variables import CommandFlagList
