+++ test/framework/type_checking.py
@@ -28,7 +28,7 @@ Unit tests for easyconfig/types.py
 @author: Kenneth Hoste (Ghent University)
 """
 import sys
-from test.framework.utilities import EnhancedTestCase
+from test.framework.utilities import EnhancedTestCase, TestLoaderFiltered
 from unittest import TextTestRunner
 
 from easybuild.framework.easyconfig.types import as_hashable, check_element_types, check_key_types, check_known_keys
@@ -38,7 +38,6 @@ from easybuild.framework.easyconfig.types import SANITY_CHECK_PATHS_DICT, STRING
 from easybuild.framework.easyconfig.types import is_value_of_type, to_name_version_dict, to_dependencies, to_dependency
 from easybuild.framework.easyconfig.types import to_list_of_strings_and_tuples, to_sanity_check_paths_dict
 from easybuild.tools.build_log import EasyBuildError
-from easybuild.tools.testfilter import TestLoaderFiltered
 
 
 class TypeCheckingTest(EnhancedTestCase):
